# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ahref Service",
    "version": "1.0.0",
    "description": "Domain-keyed cache of Ahrefs domain authority and traffic data. Keyed per domain (www folded to apex, other subdomains distinct); no outlet/campaign/brand coupling."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [],
  "components": {
    "schemas": {
      "AhrefDataType": {
        "type": "string",
        "enum": [
          "authority",
          "traffic"
        ]
      },
      "UpdateDomainRatingBody": {
        "type": "object",
        "properties": {
          "domain": {
            "type": "string"
          },
          "dataType": {
            "$ref": "#/components/schemas/AhrefDataType"
          },
          "dataCapturedAt": {
            "type": "string",
            "format": "date-time"
          },
          "urlInput": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "rawData": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            }
          },
          "authorityDomainRating": {
            "type": "integer",
            "nullable": true
          },
          "authorityUrlRating": {
            "type": "integer",
            "nullable": true
          },
          "authorityBacklinks": {
            "type": "integer",
            "nullable": true
          },
          "authorityRefdomains": {
            "type": "integer",
            "nullable": true
          },
          "authorityDofollowBacklinks": {
            "type": "integer",
            "nullable": true
          },
          "authorityDofollowRefdomains": {
            "type": "integer",
            "nullable": true
          },
          "trafficMonthlyAvg": {
            "type": "integer",
            "nullable": true
          },
          "costMonthlyAvg": {
            "type": "number",
            "nullable": true
          },
          "trafficHistory": {
            "nullable": true
          },
          "trafficTopPages": {
            "nullable": true
          },
          "trafficTopCountries": {
            "nullable": true
          },
          "trafficTopKeywords": {
            "nullable": true
          },
          "overallSearchTraffic": {
            "type": "number",
            "nullable": true
          },
          "overallSearchTrafficHistory": {
            "nullable": true
          },
          "overallSearchTrafficValue": {
            "type": "number",
            "nullable": true
          },
          "overallSearchTrafficValueHistory": {
            "nullable": true
          },
          "overallSearchTrafficByCountry": {
            "nullable": true
          },
          "trafficByCountry": {
            "nullable": true
          },
          "overallSearchTrafficKeywords": {
            "nullable": true
          }
        },
        "required": [
          "domain",
          "dataType",
          "dataCapturedAt",
          "rawData"
        ]
      },
      "DrComputeBody": {
        "type": "object",
        "properties": {
          "domains": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "minItems": 1,
            "description": "Domains to fetch DR for. Normalized server-side (www stripped, case-folded); other subdomains kept distinct."
          }
        },
        "required": [
          "domains"
        ]
      },
      "DrStatusResponse": {
        "type": "object",
        "properties": {
          "domain": {
            "type": "string"
          },
          "drToUpdate": {
            "type": "boolean"
          },
          "drUpdateReason": {
            "type": "string",
            "nullable": true
          },
          "drLatestSearchDate": {
            "type": "string",
            "nullable": true
          },
          "latestValidDr": {
            "type": "integer",
            "nullable": true
          },
          "latestValidDrDate": {
            "type": "string",
            "nullable": true
          },
          "needsUpdate": {
            "type": "boolean"
          }
        },
        "required": [
          "domain",
          "drToUpdate",
          "drUpdateReason",
          "drLatestSearchDate",
          "latestValidDr",
          "latestValidDrDate",
          "needsUpdate"
        ]
      },
      "LowDrResponse": {
        "type": "object",
        "properties": {
          "domain": {
            "type": "string"
          },
          "drToUpdate": {
            "type": "boolean"
          },
          "drUpdateReason": {
            "type": "string",
            "nullable": true
          },
          "drLatestSearchDate": {
            "type": "string",
            "nullable": true
          },
          "latestValidDr": {
            "type": "integer",
            "nullable": true
          },
          "latestValidDrDate": {
            "type": "string",
            "nullable": true
          },
          "needsUpdate": {
            "type": "boolean"
          },
          "hasLowDomainRating": {
            "type": "boolean",
            "nullable": true
          }
        },
        "required": [
          "domain",
          "drToUpdate",
          "drUpdateReason",
          "drLatestSearchDate",
          "latestValidDr",
          "latestValidDrDate",
          "needsUpdate",
          "hasLowDomainRating"
        ]
      },
      "TrafficComputeBody": {
        "type": "object",
        "properties": {
          "domains": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "minItems": 1,
            "description": "Domains to fetch traffic for. Normalized server-side (www stripped, case-folded); other subdomains kept distinct."
          }
        },
        "required": [
          "domains"
        ]
      },
      "TrafficMonthly": {
        "type": "object",
        "properties": {
          "month": {
            "type": "string",
            "description": "First day of the month (YYYY-MM-DD)."
          },
          "organicTraffic": {
            "type": "integer",
            "nullable": true
          }
        },
        "required": [
          "month",
          "organicTraffic"
        ]
      },
      "TrafficResponse": {
        "type": "object",
        "properties": {
          "domain": {
            "type": "string"
          },
          "hasData": {
            "type": "boolean"
          },
          "latestDataCapturedAt": {
            "type": "string",
            "nullable": true
          },
          "trafficMonthlyAvg": {
            "type": "integer",
            "nullable": true
          },
          "trafficValueMonthlyAvg": {
            "type": "integer",
            "nullable": true
          },
          "topPages": {
            "nullable": true
          },
          "topCountries": {
            "nullable": true
          },
          "topKeywords": {
            "nullable": true
          },
          "monthlyOrganicTraffic": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrafficMonthly"
            }
          }
        },
        "required": [
          "domain",
          "hasData",
          "latestDataCapturedAt",
          "trafficMonthlyAvg",
          "trafficValueMonthlyAvg",
          "monthlyOrganicTraffic"
        ]
      }
    },
    "parameters": {},
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok"
                      ]
                    },
                    "service": {
                      "type": "string",
                      "enum": [
                        "ahref-service"
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "service"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/orgs/domains/dr-status": {
      "get": {
        "summary": "Get DR status for a list of domains",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated domains. Normalized server-side: www stripped, case-folded; other subdomains kept distinct."
            },
            "required": true,
            "description": "Comma-separated domains. Normalized server-side: www stripped, case-folded; other subdomains kept distinct.",
            "name": "domains",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Org UUID from client-service (required)"
            },
            "required": true,
            "description": "Org UUID from client-service (required)",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "User UUID from client-service"
            },
            "required": false,
            "description": "User UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Caller's run ID from runs-service"
            },
            "required": false,
            "description": "Caller's run ID from runs-service",
            "name": "x-run-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "DR status for requested domains",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DrStatusResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/orgs/domains/dr-compute": {
      "post": {
        "summary": "Scrape Ahrefs DR for domains on demand (declares cost + authorizes)",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Org UUID from client-service (required)"
            },
            "required": true,
            "description": "Org UUID from client-service (required)",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "User UUID from client-service"
            },
            "required": false,
            "description": "User UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Caller's run ID from runs-service"
            },
            "required": false,
            "description": "Caller's run ID from runs-service",
            "name": "x-run-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DrComputeBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "DR status for the requested domains after the scrape",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DrStatusResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/orgs/domains/traffic-history": {
      "get": {
        "summary": "Get traffic (latest snapshot + monthly organic series) for domains",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated domains. Normalized server-side: www stripped, case-folded; other subdomains kept distinct."
            },
            "required": true,
            "description": "Comma-separated domains. Normalized server-side: www stripped, case-folded; other subdomains kept distinct.",
            "name": "domains",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Org UUID from client-service (required)"
            },
            "required": true,
            "description": "Org UUID from client-service (required)",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "User UUID from client-service"
            },
            "required": false,
            "description": "User UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Caller's run ID from runs-service"
            },
            "required": false,
            "description": "Caller's run ID from runs-service",
            "name": "x-run-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Traffic history for the requested domains",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrafficResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/orgs/domains/traffic-compute": {
      "post": {
        "summary": "Scrape Ahrefs traffic for domains on demand (declares cost + authorizes)",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Org UUID from client-service (required)"
            },
            "required": true,
            "description": "Org UUID from client-service (required)",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "User UUID from client-service"
            },
            "required": false,
            "description": "User UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Caller's run ID from runs-service"
            },
            "required": false,
            "description": "Caller's run ID from runs-service",
            "name": "x-run-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrafficComputeBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Traffic history for the requested domains after the scrape",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrafficResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/internal/domains/dr-stale": {
      "get": {
        "summary": "Known domains whose DR needs a refresh (platform/cron)",
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Stale DR domains",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DrStatusResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/internal/domains/low-domain-rating": {
      "get": {
        "summary": "Known domains with DR < 10 (platform/cron)",
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Low DR domains",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LowDrResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/internal/domains/domain-rating": {
      "post": {
        "summary": "Ingest scraped Ahrefs data for a domain (platform worker)",
        "security": [
          {
            "apiKey": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDomainRatingBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ahrefs data stored successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "domain": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "domain"
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}